pyproject.toml
[project]
name = "cd-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib",
    "mlflow",
    "numpy",
    "optuna>=4.8.0",
    "pandas",
    "pillow",
    "requests",
    "scikit-learn",
    "torch",
    "torchvision",
    "tqdm",
    "transformers",
    "typer",
]

[project.scripts]
cdbench = "cd_bench.cli:app"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["cd_bench*"]
